split update worker out of agent
